ia/requirements.txt
absl-py==2.1.0
asttokens==3.0.0
astunparse==1.6.3
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
# contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.17.0
Flask==3.1.0
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
huggingface-hub==0.29.2
idna==3.10
ipython==8.34.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyterlab_widgets==3.0.13
keras==3.9.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy==2.0.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.50
protobuf==5.29.3
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
simple-websocket==1.1.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.18.0
tokenizers==0.21.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
wsproto==1.2.0
